requirements.txt
streamlit
pandas
sentence-transformers
chromadb
pdfplumber
docx2txt
google-generativeai
